+++ src/Raven.Server/Documents/Handlers/CountersHandler.cs
@@ -80,7 +80,7 @@ namespace Raven.Server.Documents.Handlers
                 _list = list;
             }
 
-            protected override int ExecuteCmd(DocumentsOperationContext context)
+            protected override long ExecuteCmd(DocumentsOperationContext context)
             {
                 var countersToAdd = new SortedSet<string>(StringComparer.OrdinalIgnoreCase);
                 var countersToRemove = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
@@ -337,7 +337,7 @@ namespace Raven.Server.Documents.Handlers
                 _toDispose.Add(data);
             }
 
-            protected override int ExecuteCmd(DocumentsOperationContext context)
+            protected override long ExecuteCmd(DocumentsOperationContext context)
             {
                 if (_legacyDictionary != null)
                 {
